tvheadend: don't mix up long and int in Connect() methods

+++ xbmc/pvrclients/tvheadend/HTSPData.cpp
@@ -37,7 +37,7 @@ cHTSPData::~cHTSPData()
   Close();
 }
 
-bool cHTSPData::Open(const std::string &strHostname, unsigned int iPort, const std::string &strUsername, const std::string &strPassword, long iTimeout)
+bool cHTSPData::Open(const std::string &strHostname, unsigned int iPort, const std::string &strUsername, const std::string &strPassword, int iTimeout)
 {
   if(!m_session.Connect(strHostname, iPort, iTimeout))
   {
